chore: upgrade toolchain to 1.67.0-nightly

+++ forest/cli/src/cli/snapshot_cmd.rs
@@ -129,7 +129,7 @@ impl SnapshotCommands {
 
                 let month_string = format!("{:02}", now.month() as u8);
                 let year = now.year();
-                let day_string = format!("{:02}", now.day() as u8);
+                let day_string = format!("{:02}", now.day());
                 let chain_name = chain_get_name().await.map_err(handle_rpc_err).unwrap();
 
                 let vars = HashMap::from([
